# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Admin-Panel-Api",
        "description": "Laravel Api",
        "contact": {
            "name": "Hamed",
            "email": "h.akbarzadh@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/admin/brand/store": {
            "post": {
                "tags": [
                    "Brand"
                ],
                "summary": "Store Brand in DB",
                "description": "Store Brand in DB",
                "operationId": "brandStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "persian_name",
                                    "original_name",
                                    "logo",
                                    "status",
                                    "tags"
                                ],
                                "properties": {
                                    "persian_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Test persian_name"
                                    },
                                    "original_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Test original_name"
                                    },
                                    "logo": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Test logo"
                                    },
                                    "status": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "0 Or 1"
                                    },
                                    "tags": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Test tags"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/brand/show/{brand}": {
            "get": {
                "tags": [
                    "Brand"
                ],
                "summary": "Get Brand Detail",
                "description": "Get Brand Detail",
                "operationId": "brandshow",
                "parameters": [
                    {
                        "name": "brand",
                        "in": "path",
                        "description": "Id of brand",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/brand/update/{brand}": {
            "put": {
                "tags": [
                    "Brand"
                ],
                "summary": "Update Brand in DB",
                "description": "Update Brand in DB",
                "operationId": "brandUpdate",
                "parameters": [
                    {
                        "name": "brand",
                        "in": "path",
                        "description": "Id of Article",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "persian_name",
                                    "original_name",
                                    "logo",
                                    "status",
                                    "tags"
                                ],
                                "properties": {
                                    "persian_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Test persian_name"
                                    },
                                    "original_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Test original_name"
                                    },
                                    "logo": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Test logo"
                                    },
                                    "status": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "0 Or 1"
                                    },
                                    "tags": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Test tags"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/brand/destroy/{brand}": {
            "delete": {
                "tags": [
                    "Brand"
                ],
                "summary": "Delete Brand",
                "description": "Delete Brand",
                "operationId": "brandDestroy",
                "parameters": [
                    {
                        "name": "brand",
                        "in": "path",
                        "description": "Id of Product Category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/brand/get-brand/{number}": {
            "get": {
                "tags": [
                    "Brand"
                ],
                "summary": "Get Any Brand Detail",
                "description": "Get Any Brand Detail",
                "operationId": "brandNumber",
                "parameters": [
                    {
                        "name": "number",
                        "in": "path",
                        "description": "number",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/brand/status/{brand}": {
            "get": {
                "tags": [
                    "Brand"
                ],
                "summary": "Change Status",
                "description": "Change Status",
                "operationId": "brandStatus",
                "parameters": [
                    {
                        "name": "brand",
                        "in": "path",
                        "description": "Id of Brand",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/admin/product-category/destroy/{productCategory}": {
            "delete": {
                "tags": [
                    "Product Category"
                ],
                "summary": "Delete Product Category",
                "description": "Delete Product Category",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "productCategory",
                        "in": "path",
                        "description": "Id of Product Category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}